library_hw/library_hw/requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.8.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
manage.py==0.2.10
Markdown==3.3.4
MarkupSafe==2.0.1
mixer==7.1.2
packaging==20.9
promise==2.3
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
startapp==0.2.0.0
text-unidecode==1.3
todo==0.1
uritemplate==3.0.1
urllib3==1.26.5
